+++ public/js/ng/dashboard-view/dashboard-view.component.js
@@ -110,7 +110,7 @@
             }
 
             getPromptsWidget();
-        };
+        }
 
         function getPromptsWidget () {
             for (const promptID in vm.prompts) {