+++ server/.eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-    env: {
-        node: true,
-    },
-    plugins: ['node'],
-    extends: ['plugin:node/recommended'],
-};